+++ models/projectVo.js
@@ -6,12 +6,30 @@ module.exports = function (sequelize, dataTypes) {
       type: dataTypes.STRING(32),
       field: 'proj_id',
       primaryKey: true,
-      allowNull: false
+      allowNull: false,
+      unique: {
+        msg: 'projectId is duplicated'
+      },
+      validate: {
+        notEmpty: {
+          msg: 'projectId cannot be empty'
+        },
+        len: {
+          args: [1, 32],
+          msg: 'projectId is out of range'
+        }
+      }
     },
     description: {
       type: dataTypes.STRING(256),
       field: 'proj_description',
-      allowNull: true
+      allowNull: true,
+      validate: {
+        len: {
+          args: [0, 256],
+          msg: 'description is out of range'
+        }
+      }
     }
   }, {
     timestamps: false,  // remove createAt and updateAt attributes